requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
Pillow==9.5.0
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.12.2
imblearn==0.11.0
pandas==2.1.1
